'exchange' commands tree
Move exchange specific commands into their own subtree.

Clean up dependencies a bit.  Now that they are in a shared package, we
can use them directly, instead of passing them as arguments.  For
functions that are not actually parametrised by those dependencies, it
seems like an OK compromise.
